Set proper return type
This function was always expecting to return a `size_t` value

+++ c10/util/intrusive_ptr.h
@@ -714,7 +714,7 @@ namespace intrusive_ptr {
     return wptr.release();
   }
 
-  inline uint32_t use_count(intrusive_ptr_target* self) {
+  inline size_t use_count(intrusive_ptr_target* self) {
     auto ptr = c10::intrusive_ptr<intrusive_ptr_target>::reclaim(self);
     auto r = ptr.use_count();
     ptr.release();
